pyproject.toml
[project]
name = "fastapi-dunossauro"
version = "0.1.0"
description = ""
authors = [
    {name = "Rodrigo Alves Tenório",email = "rodten23@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.120.0,<0.121.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.2,<0.15.0)"
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev fastapi_dunossauro/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fastapi_dunossauro -vv'
post_test = 'coverage html'
